+++ src/shared/kismet/index.ts
@@ -1,2 +1,3 @@
 export * from './archetype.js'
+export * from './destructure.js'
 export * from './variables.js'