+++ src/structures/parsers.ts
@@ -1,5 +1,14 @@
-import { CustomNodesManager } from './managers/index.js'
+import {
+    CustomNodesManager,
+    TextParserOptions,
+    InputTextNodeParser,
+    InputTextSequenceParser,
+} from './managers/index.js'
 
 export const Parsers = {
     Classes: new CustomNodesManager(),
+    NodeText: InputTextNodeParser,
+    SequenceText: InputTextNodeParser,
 }
+
+export type { TextParserOptions }